+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/826>)
 - Add Searcher CLI documentation
   (<https://github.com/openvinotoolkit/datumaro/pull/838>)
+- Add version to dataset exported as datumaro format
+  (<https://github.com/openvinotoolkit/datumaro/pull/842>)
+- Add Ava action data format support
+  (<https://github.com/openvinotoolkit/datumaro/pull/847>)
 - Add YOLO Loose format
   (<https://github.com/openvinotoolkit/datumaro/pull/856>)
 
@@ -25,6 +29,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/835>)
 - Fix CIFAR10 and 100 detect function
   (<https://github.com/openvinotoolkit/datumaro/pull/836>)
+- Choose the top priority detect format for all directory depths
+  (<https://github.com/openvinotoolkit/datumaro/pull/839>)
+- Fix MVTec format detect function
+  (<https://github.com/openvinotoolkit/datumaro/pull/843>)
 
 ## 24/02/2023 - Release v1.0.0
 ### Added